rest: Authorization.

+++ routes/web.php
@@ -19,14 +19,16 @@ $router->get('/', function () use ($router) {
     return $router->app->version();
 });
 
-$router->get('users', 'UsersController@index');
-$router->get('users/{id}', 'UsersController@show');
-$router->post('users', 'UsersController@store');
-$router->put('users/{id}', 'UsersController@update');
-$router->delete('users/{id}', 'UsersController@destroy');
+/*$router->post('login', 'AuthController@login');
+
+$router->get('users', ['middleware' => 'auth', 'users' => 'UsersController@index']);
+$router->get('users/{id}', ['middleware' => 'auth', 'users' => 'UsersController@show']);
+$router->post('users', ['middleware' => 'auth', 'users' => 'UsersController@store']);
+$router->put('users/{id}', ['middleware' => 'auth', 'users' => 'UsersController@update']);
+$router->delete('users/{id}', ['middleware' => 'auth', 'users' => 'UsersController@destroy']);
 
 $router->get('sensors', 'SensorsController@index');
-$router->get('sensors/{id}', 'sensorsController@show');
+$router->get('sensors/{id}', 'SensorsController@show');
 $router->post('sensors', 'SensorsController@store');
 $router->put('sensors/{id}', 'SensorsController@update');
 $router->delete('sensors/{id}', 'SensorsController@destroy');
@@ -35,4 +37,22 @@ $router->get('actuators', 'ActuatorsController@index');
 $router->get('actuators/{id}', 'ActuatorsController@show');
 $router->post('actuators', 'ActuatorsController@store');
 $router->put('actuators/{id}', 'ActuatorsController@update');
-$router->delete('actuators/{id}', 'ActuatorsController@destroy');
+$router->delete('actuators/{id}', 'ActuatorsController@destroy');*/
+
+$router->post('login','AuthController@login');
+
+
+function recurso($router, $url, $modelo): void
+{
+    $router->get("$url",$modelo."Controller@index");
+    $router->get("$url/{id}",$modelo."Controller@show");
+    $router->post("$url",$modelo."Controller@store");
+    $router->put("$url/{id}",$modelo."Controller@update");
+    $router->delete("$url/{id}",$modelo."Controller@destroy");
+}
+
+$router->group(['middleware' => 'auth'], function () use ($router) {
+    recurso($router, 'users', 'Users');
+    recurso($router, 'sensors', 'Sensors');
+    recurso($router, 'actuators', 'Actuators');
+});
